Modify file structure
