+++ .coverage
@@ -1,1 +1,1 @@
-!coverage.py: This is a private format, don't read it directly!{"lines":{"F:\\ESSA\\2018\\week2\\challenge3\\__init__.py":[1],"F:\\ESSA\\2018\\week2\\challenge3\\test\\__init__.py":[1],"F:\\ESSA\\2018\\week2\\challenge3\\test\\test_api.py":[1,2,3,4,6,7,8,14,27,46,51,55,59,83,102,121,125,129,133,137,10,48,49,122,123,126,127,130,131,134,135,56,57,52,53,28,29,30,31,32,34,36,37,38,40,42,43,44,84,85,86,87,88,90,91,92,94,96,98,99,100,103,104,105,106,107,109,110,111,113,114,115,117,118,119,60,61,62,63,64,66,67,68,70,71,72,73,74,75,76,77,79,80,81,15,16,17,18,19,21,22,23,24,25],"F:\\ESSA\\2018\\week2\\challenge3\\api\\__init__.py":[1,2,3,4,5,6,10,11,12,13,14,17,18,19,20,23,24,25,26,27,28,29,32,34],"F:\\ESSA\\2018\\week2\\challenge3\\api\\ridesview.py":[1,2,3,4,5,8,9,12,13,16,21,40,41,44,53,54,57,80,81,84,100,101,104,115,126,18,117,118,119,120,121,46,47,59,60,61,62,63,64,65,66,67,70,71,72,73,74,86,87,88,89,91,92,93,94,95,107,108,109,110,129,130,131,132,135,136,137,138,24,25,26,27,28,29,31,32,33,34,36],"F:\\ESSA\\2018\\week2\\challenge3\\data\\__init__.py":[1],"F:\\ESSA\\2018\\week2\\challenge3\\data\\db.py":[1,4,7,19,25,9,10,11,12,13,14,16,17],"F:\\ESSA\\2018\\week2\\challenge3\\api\\models\\__init__.py":[1],"F:\\ESSA\\2018\\week2\\challenge3\\api\\models\\ride_details.py":[1,2,3,8,9,10,13,31,49,59,69,87,100,111,119,129,144,156,15,16,17,18,19,21,22,23,24,25,26,28,29,131,132,133,134,135,137,138,122,123,124,125,126,139,141,142,33,34,35,36,38,39,40,41,42,43,45,61,62,63,64,66,71,72,75,76,77,78,79,80,81,82,89,90,91,92,93,94,95,96,102,103,104,105,106,107,113,115,116,117,146,147,148,149,150,151,152,158,159,160,161,162,51,52,55,56,57],"F:\\ESSA\\2018\\week2\\challenge3\\run.py":[],"F:\\ESSA\\2018\\week2\\challenge3\\data\\tables.py":[]}}
+!coverage.py: This is a private format, don't read it directly!{"lines":{"H:\\RideEssa\\api\\views\\ridesview.py":[1,2,4,5,8,11,12,13,17,35,36,39,47,48,50,72,73,75,91,92,94,104,114,52,53,54,55,56,57,58,59,60,63,64,65,66,67,20,21,22,23,24,25,27,28,29,30,32,15,97,98,99,100,106,107,108,109,110,41,42,77,78,79,80,82,83,84,85,86,118,119,120,121,122,123,126,127,128,129],"H:\\RideEssa\\api\\database\\db.py":[1,2,5,6,27,33,37,42,67,8,9,13,14,15,20,21,23,43,49,51,56,58,64,38,39,40],"H:\\RideEssa\\api\\models\\ride_details.py":[1,3,4,5,8,9,11,30,49,69,80,98,112,124,133,144,160,173,71,72,73,74,75,77,82,83,84,87,88,89,90,91,92,93,94,51,52,53,54,57,58,59,60,61,62,63,13,14,15,16,17,18,20,21,22,23,24,25,27,28,114,115,116,117,118,119,120,126,127,129,130,131,146,147,148,149,150,151,153,154,136,137,138,139,140,141,155,157,158,32,33,34,35,36,38,39,40,41,42,43,45,100,101,102,103,104,105,106,107,108,162,163,164,165,166,167,168,169,175,176,177,178,179,180]}}
