+++ .gitignore
@@ -1,5 +1,4 @@
-.vscode
-__pycache__
-.pytest_cache
-.idea
+*.pyc
+.vscode/*
+!.vscode/settings.json
 htmlcov
