+++ api/__init__.py
@@ -1,23 +1,27 @@
-"""This contains the libraries used and the endpoints"""
-from flask import Flask 
-from flask import request
-from flask import jsonify
-from flask_restful import Resource, Api
-from flask_jwt_extended import (
-    JWTManager, jwt_required, create_access_token,
-    get_jwt_identity
-)
-from api.ridesview import Ride
-from api.ridesview import RideOffer
-from api.ridesview import Register
-from api.ridesview import Login
-from api.ridesview import MakeRequest
+from api.views.ridesview import Ride
+from api.views.ridesview import RideOffer
+from api.views.ridesview import Register
+from api.views.ridesview import Login
+from api.views.ridesview import MakeRequest
 
+from flask import Flask
+from flask import jsonify
+from flask_restful import Api
+from flask_jwt_extended import JWTManager
 
 app = Flask(__name__)
 app.config["JWT_SECRET_KEY"] = "GHFVJDFJKSNJKDVNKJSDNKVJJJNDJK"
 jwt = JWTManager(app)
 api = Api(app)
+
+app.config["DEBUG"] = True
+app.config["TESTING"] = False
+app.config.from_object(__name__)
+
+
+@app.errorhandler(404)
+def not_found(error):
+    return jsonify({"message": error.description}), 404
 
 
 api.add_resource(Ride, "/rides", "/users/rides")
@@ -29,6 +33,4 @@ api.add_resource(MakeRequest, "/users/rides/<int:ride_id>/requests",
                  "/users/rides/<int:ride_id>/requests/<int:request_id>")
 
 
-@app.errorhandler(404)
-def not_found(error):
-    return jsonify({"message":error.description}), 404
+
