+++ run.py
@@ -1,5 +1,4 @@
-""" this runs the server""" 
 from api import app
 
 if __name__ == "__main__":
-    app.run(port=5000, debug=True)
+    app.run()
